Adds minor fixes and renames files. Creates API folder and moves API specific files in here
