+++ JavaScript/modules/API/deletePosts.mjs
@@ -1,11 +1,11 @@
-import { tryCatchError } from "./error.mjs";
-import { preventFormRefresh } from "./noRefresh.mjs";
-import * as consts from "./consts.mjs";
-import { API_BASE_URL } from "./inputs.mjs";
+import { tryCatchError } from "../error.mjs";
+import { preventFormRefresh } from "../noRefresh.mjs";
+import * as consts from "../consts.mjs";
+import { API_BASE_URL, authorEndpoint, socialEndpoint } from "../inputs.mjs";
 
-const queryString = document.location.search;
+/*const queryString = document.location.search;
 const parameters = new URLSearchParams(queryString);
-const id = parameters.get("id");
+const id = parameters.get("id");*/
 
 export async function deletePost(url, event) {
   try {
@@ -17,14 +17,14 @@ export async function deletePost(url, event) {
       },
     };
     const response = await fetch(
-      API_BASE_URL + "/api/v1/social/posts/" + id + "?_author=true",
+      API_BASE_URL + socialEndpoint + consts.id + authorEndpoint,
       getData
     );
     if (response.ok) {
       alert("Post has been deleted");
       setTimeout(() => {
         window.location.href = "/feed/index.html";
-      }, 1200);
+      }, 500);
     } else {
       alert("You can't delete other users post!");
     }
