+++ feed/single-post/singlePost.js
@@ -1,10 +1,12 @@
-import { API_BASE_URL } from "../../JavaScript/modules/inputs.mjs";
+import {
+  API_BASE_URL,
+  authorEndpoint,
+  socialEndpoint,
+} from "../../JavaScript/modules/inputs.mjs";
 import * as consts from "../../JavaScript/modules/consts.mjs";
 import { displaySinglePost } from "../../JavaScript/modules/displayFunction.mjs";
-import { deletePost } from "../../JavaScript/modules/deletePosts.mjs";
-import { updatePost } from "../../JavaScript/modules/API/update.mjs";
-
-//Query params to fetch ID
+import { deletePost } from "../../JavaScript/modules/API/deletePosts.mjs";
+import { updatePost } from "../../JavaScript/modules/API/updatePosts.mjs";
 
 const form = document.getElementById("formEdit");
 
@@ -12,7 +14,7 @@ console.log(consts.id);
 
 async function fillFormBeforeUpdate() {
   const post = await displaySinglePost(
-    API_BASE_URL + "/api/v1/social/posts/" + consts.id + "?_author=true"
+    API_BASE_URL + socialEndpoint + consts.id + authorEndpoint
   );
 
   console.log(form);
