+++ register/index.html
@@ -37,7 +37,7 @@
     <script
       type="module"
       defer
-      src="/JavaScript/modules/fetchContent.mjs"
+      src="/JavaScript/modules/API/fetchContent.mjs"
     ></script>
 
     <title>Musync | Register User</title>
